Move selection-commands.ts to be in the vscode directory
